Move even more things to asset processor
