+++ scraper/src/mindtouch2zim/utils.py
@@ -1,6 +1,9 @@
 from pathlib import Path
 from urllib.parse import urlparse
 
+from zimscraperlib.zim import Creator
+from zimscraperlib.zim.indexing import IndexData
+
 
 def get_asset_path_from_url(online_url: str, already_used_paths: list[Path]) -> Path:
     """Computes the path where one should store its asset based on its online URL
@@ -52,3 +55,39 @@ def is_better_srcset_descriptor(
     if current_best_descriptor[-1:] != new_descriptor[-1:]:
         return False
     return int(new_descriptor[:-1]) > int(current_best_descriptor[:-1])
+
+
+def add_item_for(
+    creator: Creator,
+    path: str,
+    title: str | None = None,
+    *,
+    fpath: Path | None = None,
+    content: bytes | str | None = None,
+    mimetype: str | None = None,
+    is_front: bool | None = None,
+    should_compress: bool | None = None,
+    delete_fpath: bool | None = False,
+    duplicate_ok: bool | None = None,
+    index_data: IndexData | None = None,
+    auto_index: bool = True,
+):
+    """
+    Boilerplate to avoid repeating pyright ignore
+
+    To be removed, once upstream issue is solved, see
+    https://github.com/openzim/libretexts/issues/26
+    """
+    creator.add_item_for(  # pyright: ignore[reportUnknownMemberType]
+        path=path,
+        title=title,
+        fpath=fpath,
+        content=content,
+        mimetype=mimetype,
+        is_front=is_front,
+        should_compress=should_compress,
+        delete_fpath=delete_fpath,
+        duplicate_ok=duplicate_ok,
+        index_data=index_data,
+        auto_index=auto_index,
+    )
